oldbkp/requirements.txt
djitellopy
opencv-python
pynput
numpy
pyserial
scikit-learn
matplotlib
